try now

+++ internal/flags/flags_test.go
@@ -27,8 +27,8 @@ type MockPTHICommands struct{}
 func (c MockPTHICommands) Open(bool) error {
 	return nil
 }
-func (c MockPTHICommands) GetIsAMTEnabled() (state int, err error) {
-	return 0, nil
+func (c MockPTHICommands) GetIsAMTEnabled() (state uint8, err error) {
+	return uint8(0), nil
 }
 func (c MockPTHICommands) Close() {}
